Hack to execute test cases
Added resources files with the public keys as these keys are too
long for execution.
Currently, it is a hack as for both cases, the resources read
function is hard coded.
Further, for the test case verifyingIIPTest the change does not
work. A following function throws an exception. Need to check that.